+++ src/main.cpp
@@ -73,6 +73,19 @@ bool avx2_enabled() {
 #endif
 }
 
+InputBuffer get_input_buffer(const CommandLineOptions& opt) {
+    if (opt.is_SE) {
+        return InputBuffer(opt.reads_filename1, "", opt.chunk_size, false);
+    } else if (opt.is_interleaved) {
+        if (opt.reads_filename2 != "") {
+            throw BadParameter("Cannot specify both --interleaved and specify two read files");
+        }
+        return InputBuffer(opt.reads_filename1, "", opt.chunk_size, true);
+    } else {
+        return InputBuffer(opt.reads_filename1, opt.reads_filename2, opt.chunk_size, false);
+    }
+}
+
 int run_strobealign(int argc, char **argv) {
     auto opt = parse_command_line_arguments(argc, argv);
 
@@ -211,44 +224,20 @@ int run_strobealign(int argc, char **argv) {
 
     logger.info() << "Running in " << (opt.is_SE ? "single-end" : "paired-end") << " mode" << std::endl;
 
-    if (opt.is_SE) {
-        auto ks = open_fastq(opt.reads_filename1);
+    InputBuffer input_buffer = get_input_buffer(opt);
+    OutputBuffer output_buffer(out);
 
-        InputBuffer input_buffer(ks, ks, opt.chunk_size);
-        OutputBuffer output_buffer(out);
-
-        std::vector<std::thread> workers;
-        for (int i = 0; i < opt.n_threads; ++i) {
-            std::thread consumer(perform_task_SE, std::ref(input_buffer), std::ref(output_buffer),
-                std::ref(log_stats_vec[i]), std::ref(aln_params),
-                std::ref(map_param), std::ref(index_parameters), std::ref(references),
-                std::ref(index), std::ref(opt.read_group_id));
-            workers.push_back(std::move(consumer));
-        }
-
-        for (size_t i = 0; i < workers.size(); ++i) {
-            workers[i].join();
-        }
+    std::vector<std::thread> workers;
+    for (int i = 0; i < opt.n_threads; ++i) {
+        std::thread consumer(perform_task, std::ref(input_buffer), std::ref(output_buffer),
+            std::ref(log_stats_vec[i]), std::ref(aln_params),
+            std::ref(map_param), std::ref(index_parameters), std::ref(references),
+            std::ref(index), std::ref(opt.read_group_id));
+        workers.push_back(std::move(consumer));
     }
-    else {
-        auto ks1 = open_fastq(opt.reads_filename1);
-        auto ks2 = open_fastq(opt.reads_filename2);
 
-        InputBuffer input_buffer(ks1, ks2, opt.chunk_size);
-        OutputBuffer output_buffer(out);
-
-        std::vector<std::thread> workers;
-        for (int i = 0; i < opt.n_threads; ++i) {
-            std::thread consumer(perform_task_PE, std::ref(input_buffer), std::ref(output_buffer),
-                std::ref(log_stats_vec[i]), std::ref(aln_params),
-                std::ref(map_param), std::ref(index_parameters), std::ref(references),
-                std::ref(index), std::ref(opt.read_group_id));
-            workers.push_back(std::move(consumer));
-        }
-
-        for (size_t i = 0; i < workers.size(); ++i) {
-            workers[i].join();
-        }
+    for (size_t i = 0; i < workers.size(); ++i) {
+        workers[i].join();
     }
     logger.info() << "Done!\n";